add custom hooks for slider
